+++ src/main/java/ru/shummi/operation/impl/AccountTransferHandler.java
@@ -33,8 +33,8 @@ public class AccountTransferHandler implements OperationHandler {
 
     @Override
     public String perform() {
-        int sourceAccountId = ioHandler.getInteger("Enter source account ID");
-        int targetAccountId = ioHandler.getInteger("Enter target account ID");
+        long sourceAccountId = ioHandler.getLong("Enter source account ID");
+        long targetAccountId = ioHandler.getLong("Enter target account ID");
         String amount = ioHandler.getString("Enter amount to transfer");
         BigDecimal transfer = new BigDecimal(amount);
         userAccountService.transferAccountFromIdToId(sourceAccountId, targetAccountId, transfer);
