+++ src/main/java/ru/shummi/operation/impl/AccountWithdrawHandler.java
@@ -30,7 +30,7 @@ public class AccountWithdrawHandler implements OperationHandler {
 
     @Override
     public String perform() {
-        int accountId = ioHandler.getInteger("Enter account ID to withdraw from");
+        long accountId = ioHandler.getLong("Enter account ID to withdraw from");
         String amount = ioHandler.getString("Enter amount to withdraw");
         BigDecimal money = new BigDecimal(amount);
         userAccountService.withdrawAccountById(accountId, money);
